changed naming convention of files/folders. added content to SignUp.jsx
